test: complete separation / perfect prediction
The test checks that the logreg methods return objects of the same class in a well-behaved case and in a perfect separation case.
The two cases differ only by exclusion / inclusion of a perfect predictor of a dichotomous DV.
This means we are testing discrepancy in behaviour of the same mice.impute.method function due to perfect prediction.
This means that:
- if the well-behaved case returns a factor (expected outcome), while the perfect separation case returns an error or warning, the test fails (glm returns warning in case of perfect separation)
- if both return a factor, the test passes.
- if there is a code breaking bug that makes both cases return an error, the test does not fail! This is desirable because it does not point to
perfect prediction as a problem to solve. The root of the problem is elsewhere.