fix: adapt x provided to cv.glmnet to include vector of 1s
This allows the edge case where the x argument of the methods has only 1 predictor to run.
The presence of a vector of 1s in the matrix provided to the cv.glmnet function does not impact in any way its performance.
cv.glmnet() will shrink reg coefs for constants to 0.